backing out config move
