+++ source/extensions/all_extensions.bzl
@@ -6,6 +6,7 @@ load("@envoy_build_config//:extensions_build_config.bzl", "EXTENSIONS")
 _required_extensions = {
     "envoy.http.original_ip_detection.xff": "//source/extensions/http/original_ip_detection/xff:config",
     "envoy.request_id.uuid": "//source/extensions/request_id/uuid:config",
+    "envoy.transport_sockets.tls": "//source/extensions/transport_sockets/tls:config",
     # To provide default round robin load balancer.
     "envoy.load_balancing_policies.round_robin": "//source/extensions/load_balancing_policies/round_robin:config",
 }
