+++ test/common/listener_manager/BUILD
@@ -80,7 +80,6 @@ envoy_cc_test(
         "//source/common/network:utility_lib",
         "//source/common/protobuf",
         "//source/common/router:string_accessor_lib",
-        "//source/common/tls:config",
         "//source/common/tls:ssl_socket_lib",
         "//source/extensions/filters/listener/original_dst:config",
         "//source/extensions/filters/listener/proxy_protocol:config",
@@ -91,6 +90,7 @@ envoy_cc_test(
         "//source/extensions/matching/network/common:inputs_lib",
         "//source/extensions/request_id/uuid:config",
         "//source/extensions/transport_sockets/raw_buffer:config",
+        "//source/extensions/transport_sockets/tls:config",
         "//test/integration/filters:test_listener_filter_lib",
         "//test/server:utility_lib",
         "//test/test_common:network_utility_lib",
@@ -108,12 +108,12 @@ envoy_cc_test(
     tags = ["nofips"],
     deps = [
         ":listener_manager_impl_test_lib",
-        "//source/common/tls:config",
         "//source/extensions/filters/http/router:config",
         "//source/extensions/matching/network/common:inputs_lib",
         "//source/extensions/quic/server_preferred_address:fixed_server_preferred_address_config_factory_config",
         "//source/extensions/request_id/uuid:config",
         "//source/extensions/transport_sockets/raw_buffer:config",
+        "//source/extensions/transport_sockets/tls:config",
         "//test/integration/filters:test_listener_filter_lib",
         "//test/integration/filters:test_network_filter_lib",
         "//test/server:utility_lib",
@@ -137,11 +137,11 @@ envoy_cc_test(
         "//source/common/network:socket_option_lib",
         "//source/common/network:utility_lib",
         "//source/common/protobuf",
-        "//source/common/tls:config",
         "//source/common/tls:ssl_socket_lib",
         "//source/extensions/filters/network/http_connection_manager:config",
         "//source/extensions/matching/network/common:inputs_lib",
         "//source/extensions/transport_sockets/raw_buffer:config",
+        "//source/extensions/transport_sockets/tls:config",
         "//test/mocks/network:network_mocks",
         "//test/mocks/server:drain_manager_mocks",
         "//test/mocks/server:factory_context_mocks",
@@ -193,7 +193,7 @@ envoy_cc_benchmark_binary(
         "//test/mocks/server:factory_context_mocks",
         "//test/mocks/stream_info:stream_info_mocks",
         # tranport socket config registration
-        "//source/common/tls:config",
+        "//source/extensions/transport_sockets/tls:config",
     ],
 )
