+++ test/common/upstream/BUILD
@@ -92,7 +92,6 @@ envoy_cc_test(
     deps = [
         ":test_cluster_manager",
         "//source/common/router:context_lib",
-        "//source/common/tls:config",
         "//source/common/upstream:load_balancer_factory_base_lib",
         "//source/extensions/clusters/eds:eds_lib",
         "//source/extensions/clusters/logical_dns:logical_dns_cluster_lib",
@@ -111,6 +110,7 @@ envoy_cc_test(
         "//source/extensions/load_balancing_policies/round_robin:config",
         "//source/extensions/load_balancing_policies/subset:config",
         "//source/extensions/network/dns_resolver/cares:config",
+        "//source/extensions/transport_sockets/tls:config",
         "//source/extensions/upstreams/http/generic:config",
         "//test/config:v2_link_hacks",
         "//test/integration/load_balancers:custom_lb_policy",
@@ -541,7 +541,7 @@ envoy_cc_test(
         "//source/extensions/clusters/strict_dns:strict_dns_cluster_lib",
         "//source/extensions/transport_sockets/raw_buffer:config",
         "//source/extensions/transport_sockets/http_11_proxy:upstream_config",
-        "//source/common/tls:config",
+        "//source/extensions/transport_sockets/tls:config",
         "//source/extensions/upstreams/http:config",
         "//source/extensions/upstreams/tcp:config",
         "//source/server:transport_socket_config_lib",
